# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatOTP - WhatsApp OTP Verification SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/current:
    get:
      operationId: getCurrentSubscription
      tags: [subscriptions]
      summary: Get current user subscription
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: No active subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions:
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/{id}/cancel:
    post:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # API Keys
...
  /payments/initialize:
    post:
      operationId: initializePayment
      tags: [payments]
      summary: Initialize a Paystack payment for a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitInput"
      responses:
        "200":
          description: Payment initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify Paystack payment and activate subscription
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: Payment not successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # User Custom WhatsApp Number
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Plan:
      type: object
      required: [id, name, period, price, currency, otpLimit, features]
      properties:
        id:
          type: integer
        name:
          type: string
        period:
          type: string
          enum: [weekly, monthly, yearly]
        price:
          type: number
        currency:
          type: string
        otpLimit:
          type: integer
        features:
          type: array
          items:
            type: string
        popular:
          type: boolean
        allowCustomNumber:
          type: boolean
        allowUsaNumbers:
          type: boolean

    # Subscriptions
    Subscription:
      type: object
      required: [id, userId, planId, status, startDate, endDate]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        plan:
          $ref: "#/components/schemas/Plan"
        status:
          type: string
          enum: [active, cancelled, expired]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        otpUsed:
          type: integer
    SubscriptionInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: integer

    # API Keys
...
    PaymentInitInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: integer
    PaymentInitResponse:
      type: object
      required: [authorizationUrl, reference, planId, amount, currency]
      properties:
        authorizationUrl:
          type: string
        reference:
          type: string
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        currency:
          type: string

    # User custom number input
...